# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Snap is an open telemetry framework designed to simplify the collection,\nprocessing and publishing of system data through a single API.\n\nThis specification should demonstrate a fully compliant swagger 2.0 spec for Snap.\nIt lists all the possible Snap Restful APIs that are available for use,\ntesting or generating spec compatible clients.",
    "title": "Snap Restful APIs.",
    "termsOfService": "There are no TOS at this moment, use at your own risk we take no responsibility.",
    "contact": {
      "name": "Snap Maintainer",
      "url": "https://github.com/micruzz82/snap",
      "email": "snap@micruzz82.slack.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/"
    },
    "version": "2.0"
  },
  "host": "127.0.0.1:8181",
  "basePath": "/v2",
  "paths": {
...
    "/plugins/{ptype}/{pname}/{pversion}/config": {
      "get": {
        "description": "An empty config is returned if there are no configs for the plugin.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "plugins"
        ],
        "summary": "Get Config",
        "operationId": "getPluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          }
        }
      },
      "put": {
        "description": "A config is JSON. For example: {\"user\":\"snap\", \"host\":\"ocean_eleven\"}.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "plugins"
        ],
        "summary": "Set Config",
        "operationId": "setPluginConfigItem",
        "parameters": [
          {
            "x-go-name": "Config",
            "name": "config",
            "in": "body",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          }
        }
      },
      "delete": {
        "description": "A minimum of one config key is required for this operation.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "plugins"
        ],
        "summary": "Delete Config",
        "operationId": "deletePluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Config",
            "name": "config",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "description": "An empty list returns if no tasks exist.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Get All",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "$ref": "#/responses/TasksResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          }
        }
      },
      "post": {
        "description": "A string representation of Snap task manifest is required.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Add",
        "operationId": "addTask",
        "parameters": [
          {
            "x-go-name": "Task",
            "description": "Create a task.",
            "name": "task",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/TaskResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "description": "The task ID is required.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Get",
        "operationId": "getTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "description": "The task ID is required.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Enable/Start/Stop",
        "operationId": "updateTaskState",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Action",
            "description": "Update the state of a task",
            "name": "action",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/TaskResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "The task ID is required.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Remove",
        "operationId": "removeTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/TaskResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/TaskErrorResponse"
          }
        }
      }
    },
    "/tasks/{id}/watch": {
      "get": {
        "description": "The task ID is required.",
        "produces": [
          "text/event-stream"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Watch",
        "operationId": "watchTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskWatchResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "CollectWorkflowMapNode": {
      "type": "object",
      "title": "CollectWorkflowMapNode represents Snap workflow data model.",
      "required": [
        "metrics"
      ],
      "properties": {
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "x-go-name": "Config"
        },
        "metrics": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/metricInfo"
          },
          "x-go-name": "Metrics"
        },
        "process": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessWorkflowMapNode"
          },
          "x-go-name": "Process"
        },
        "publish": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PublishWorkflowMapNode"
          },
          "x-go-name": "Publish"
        },
        "tags": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "x-go-name": "Tags"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/scheduler/wmap"
    },
    "ConfigDataNode": {
      "description": "Represents a set of configuration data",
      "type": "object",
      "x-go-package": "github.com/micruzz82/snap/core/cdata"
    },
...
    "Error": {
      "description": "Unsuccessful generic response to a failed API call",
      "type": "object",
      "properties": {
        "fields": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Fields"
        },
        "message": {
          "type": "string",
          "x-go-name": "ErrorMessage"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
...
    "ProcessWorkflowMapNode": {
      "type": "object",
      "required": [
        "plugin_name"
      ],
      "properties": {
        "config": {
          "description": "Config the configuration of a processor.",
          "type": "object",
          "additionalProperties": {
            "type": "object"
          },
          "x-go-name": "Config"
        },
        "plugin_name": {
          "type": "string",
          "x-go-name": "PluginName"
        },
        "plugin_version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PluginVersion"
        },
        "process": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessWorkflowMapNode"
          },
          "x-go-name": "Process"
        },
        "publish": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PublishWorkflowMapNode"
          },
          "x-go-name": "Publish"
        },
        "target": {
          "type": "string",
          "x-go-name": "Target"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/scheduler/wmap"
    },
    "PublishWorkflowMapNode": {
      "type": "object",
      "required": [
        "plugin_name",
        "config"
      ],
      "properties": {
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          },
          "x-go-name": "Config"
        },
        "plugin_name": {
          "type": "string",
          "x-go-name": "PluginName"
        },
        "plugin_version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PluginVersion"
        },
        "target": {
          "type": "string",
          "x-go-name": "Target"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/scheduler/wmap"
    },
...
    "Schedule": {
      "type": "object",
      "title": "Schedule defines a scheduler.",
      "required": [
        "type",
        "interval"
      ],
      "properties": {
        "count": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Count"
        },
        "interval": {
          "type": "string",
          "x-go-name": "Interval"
        },
        "start_timestamp": {
          "x-go-name": "StartTimestamp"
        },
        "stop_timestamp": {
          "x-go-name": "StopTimestamp"
        },
        "type": {
          "type": "string",
          "enum": [
            "simple",
            " windowed",
            " streaming",
            " cron"
          ],
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/core"
    },
    "StreamedMetric": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "x-go-name": "Data"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "tags": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Tags"
        },
        "timestamp": {
          "x-go-name": "Timestamp"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "StreamedMetrics": {
      "type": "array",
      "title": "StreamedMetrics defines a slice of streamed metrics.",
      "items": {
        "$ref": "#/definitions/StreamedMetric"
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "StreamedTaskEvent": {
      "type": "object",
      "title": "StreamedTaskEvent defines the task watching data type.",
      "properties": {
        "event": {
          "$ref": "#/definitions/StreamedMetrics"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "type": {
          "type": "string",
          "x-go-name": "EventType"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "Task": {
      "type": "object",
      "title": "Task represents Snap task definition.",
      "properties": {
        "creation_timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreationTimestamp"
        },
        "deadline": {
          "type": "string",
          "x-go-name": "Deadline"
        },
        "failed_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "FailedCount"
        },
        "hit_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "HitCount"
        },
        "href": {
          "type": "string",
          "x-go-name": "Href"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "last_failure_message": {
          "type": "string",
          "x-go-name": "LastFailureMessage"
        },
        "last_run_timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LastRunTimestamp"
        },
        "max-failures": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "MaxFailures"
        },
        "miss_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "MissCount"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "schedule": {
          "$ref": "#/definitions/Schedule"
        },
        "start": {
          "type": "boolean",
          "x-go-name": "Start"
        },
        "task_state": {
          "type": "string",
          "x-go-name": "TaskState"
        },
        "version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version"
        },
        "workflow": {
          "$ref": "#/definitions/WorkflowMap"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "Tasks": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Task"
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "UnauthError": {
      "type": "object",
      "title": "UnauthError defines the error type of an unauthorized response.",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/mgmt/rest/v2"
    },
    "WorkflowMap": {
      "description": "WorkflowMap represents a map of a desired workflow that is used to create a scheduleWorkflow",
      "type": "object",
      "required": [
        "collect"
      ],
      "properties": {
        "collect": {
          "$ref": "#/definitions/CollectWorkflowMapNode"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/scheduler/wmap"
    },
    "metricInfo": {
      "type": "object",
      "properties": {
        "version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version_"
        }
      },
      "x-go-package": "github.com/micruzz82/snap/scheduler/wmap"
    }
  },
  "responses": {
    "ErrorResponse": {
      "description": "ErrorResponse represents the Snap error response type.\n\nIt includes an error message and a map of fields.",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
...
    "PluginConfigResponse": {
      "description": "PluginConfigResponse represents the response of a plugin config items.",
      "schema": {
        "$ref": "#/definitions/ConfigDataNode"
      }
    },
...
    "TaskErrorResponse": {
      "description": "TaskErrorResponse returns removing a task error."
    },
    "TaskResponse": {
      "description": "TaskResponse returns a task.",
      "schema": {
        "$ref": "#/definitions/Task"
      }
    },
    "TaskWatchResponse": {
      "description": "TaskWatchResponse defines the response of the task watching stream.",
      "schema": {
        "type": "object",
        "properties": {
          "task_watch": {
            "$ref": "#/definitions/StreamedTaskEvent"
          }
        }
      }
    },
    "TasksResponse": {
      "description": "TasksResponse returns a list of created tasks.",
      "schema": {
        "type": "object",
        "properties": {
          "tasks": {
            "$ref": "#/definitions/Tasks"
          }
        }
      }
    },
    "UnauthResponse": {
      "description": "UnauthResponse returns Unauthorized error struct message.",
      "schema": {
        "$ref": "#/definitions/UnauthError"
      }
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "description": "HTTP basic authentication.",
      "type": "basic"
    }
  },
  "security": [
    {
      "basicAuth": [
        "[]"
      ]
    }
  ]
}